feat(silent_admin_change): extract secure module to separate file and add tests
- Move inline pub mod secure { ... } to src/secure.rs per CONTRIBUTING.md §8
- Remove unused symbol_short import from lib.rs
- Add 3 secure-side tests: emits event, requires auth, reinit panics
- Add docs/vulnerabilities.md entry for silent_admin_change
